requirements.txt
affine==2.4.0
aiobotocore==2.12.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
asciitree==0.3.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blosc2==2.5.1
botocore==1.34.51
bounded-pool-executor==0.0.3
cachetools==5.3.3
Cartopy==0.22.0
certifi==2024.2.2
cfchecker==4.1.0
cffi==1.16.0
cftime==1.6.3
cfunits==3.3.6
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudnetpy==1.60.3
cloudnetpy_qc==1.17.1
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
dask==2024.3.1
debugpy==1.8.1
decorator==5.1.1
earthaccess==0.9.0
ephem==4.1.5
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.19
fiona==1.9.6
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gcsfs==2024.3.1
geopandas==0.14.3
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
h3==3.7.7
h5netcdf==1.3.0
h5py==3.10.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.22.2
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
locket==1.0.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
MetPy==1.6.1
msgpack==1.0.8
multidict==6.0.5
multimethod==1.11.2
mwrpy==0.8.3
ndindex==1.8
nest-asyncio==1.6.0
netCDF4==1.6.5
networkx==3.2.1
numba==0.59.1
numcodecs==0.12.1
numpy==1.26.4
oauthlib==3.2.2
ocf-blosc2==0.0.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pillow==10.2.0
Pint==0.23
platformdirs==4.2.0
pooch==1.8.1
pqdm==0.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pvlib==0.10.4
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
python-cmr==0.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rasterio==1.3.9
requests==2.31.0
requests-oauthlib==2.0.0
rioxarray==0.15.1
rpgPy==0.15.1
rsa==4.9
s3fs==2024.3.1
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.3
six==1.16.0
snuggs==1.4.7
stack-data==0.6.3
tifffile==2024.2.12
timezonefinder==6.5.0
tinynetrc==1.3.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.0.7
jinja2
wcwidth==0.2.13
wget==3.2
wrapt==1.16.0
xarray==2024.2.0
yarl==1.9.4
zarr==2.17.1
zipp==3.18.1
